+++ pyproject.toml
@@ -7,7 +7,7 @@ name = "zaptools"
 description = "A toolkit for Event-Driven websocket management"
 readme = "README.md"
 authors = [{name = "Nathan Mejia", email = "nathandraco22@gmail.com"}]
-version = "0.4.6"
+version = "0.5.0"
 requires-python = ">=3.10"
 classifiers = [
     'Development Status :: 4 - Beta',
