+++ zaptools/__init__.py
@@ -1,4 +1,4 @@
-from .tools import EventContext, EventRegister
+from .core import EventContext, EventRegister
 from .connection_router import ConnectionRouter
 
 __all__ = [
